Money stuff (hotel rooms cost and make money)

+++ src/com/snake/ld31/DataContainer.java
@@ -10,7 +10,7 @@ public class DataContainer
 	
 	public static Room[][] rooms;
 	
-	public static int money;
+	public static double money;
 	
 	public static double hours;
